refactor : modularize current project
