--- src/app/app.css
@@ -1,75 +1,0 @@
-.pokemon-container {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
-  gap: 16px;
-  padding: 0;
-  margin: 0;
-  list-style-type: none;
-}
-
-.pokemon-container ul {
-  padding: 0;
-  margin: 0;
-  list-style-type: none;
-}
-
-.pokemon-card {
-  padding: 20px;
-  border: 1px solid #ddd;
-  border-radius: 8px; /* Rounded corners */
-  text-align: center;
-  text-transform: capitalize; /* Makes the first letter uppercase */
-  width: 150px;
-  cursor: pointer;
-}
-
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.7);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
-
-.modal-content {
-  background-color: white;
-  padding: 30px;
-  border-radius: 10px;
-  text-align: center;
-}
-
-.types-container {
-  margin-top: 15px;
-}
-
-.type-badge {
-  display: inline-block;
-  padding: 5px 10px;
-  border-radius: 15px;
-  background-color: #eee;
-  margin: 0 5px;
-  text-transform: capitalize;
-}
-
-.modal-header {
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-.modal-body {
-  display: flex;
-  gap: 30px; /* Creates space between the columns */
-}
-
-.left-column, .right-column {
-  flex: 1; /* This tells both columns to share the space equally */
-}
-
-.left-column img {
-  width: 100%; /* Make the image fit its column */
-}
