+++ src/app/app.routes.ts
@@ -1,3 +1,12 @@
 import { Routes } from '@angular/router';
 
-export const routes: Routes = [];
+export const routes: Routes = [
+  {
+    path: 'landing-page',
+    loadComponent: () => import('./pokedex/pages/landing-page/landing-page'),
+  },
+  {
+    path: '**',
+    redirectTo: 'landing-page',
+  },
+];
