+++ src/app/pokedex/interfaces/pokemon.model.ts
